Fix windows tests
